requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file

# Add PyTorch repository for CPU builds
--extra-index-url https://download.pytorch.org/whl/cpu

# Core framework
Django==5.1.2
gunicorn==23.0.0

# Database
psycopg2-binary==2.9.9
asyncpg==0.30.0

# Data & analysis
numpy==2.1.3
matplotlib==3.9.2
plotly==5.24.1
pytz==2025.1

# Research APIs & scraping
scholarly==1.7.11
arxiv==1.4.8
selenium==4.25.0
habanero==1.2.6
pyalex==0.18


# Messaging / async processing
kafka-python==2.0.2

# Front-end templating
Jinja2==3.1.4

#NLP & Keywords Extraction
rake-nltk==1.0.5
nltk==3.9.1
keybert==0.8.5
sentence-transformers==3.2.1
scikit-learn==1.5.2

# Hugging Face summarization
transformers==4.45.2
torch==2.6.0+cpu 
sentencepiece==0.2.0

whitenoise
aiokafka==0.10.0
httpx==0.27.0

tqdm==4.66.5

python-dotenv==1.0.1

